feat: add support for running inside a container
The cli can now work properly inside a container if the `DOCKER_CONTAINER` environment variable is
set. A devcontainer is set up for this repository both as an example and as a development
environment for the cli and libdragon itself.

+++ modules/actions/destroy.js
@@ -3,13 +3,19 @@ const path = require('path');
 
 const { destroyContainer } = require('./utils');
 const { CONFIG_FILE, LIBDRAGON_PROJECT_MANIFEST } = require('../constants');
-const { fileExists, dirExists, log } = require('../helpers');
+const { fileExists, dirExists, log, ValidationError } = require('../helpers');
 const chalk = require('chalk');
 
 /**
  * @param {import('../project-info').LibdragonInfo} libdragonInfo
  */
 const destroy = async (libdragonInfo) => {
+  if (process.env.DOCKER_CONTAINER) {
+    throw new ValidationError(
+      `Not possible to destroy the container from inside.`
+    );
+  }
+
   await destroyContainer(libdragonInfo);
 
   const projectPath = path.join(libdragonInfo.root, LIBDRAGON_PROJECT_MANIFEST);
